Add files via upload
We have a whole bunch of cars going across a super busy multi-lane highway

and our turtle is the player who has to cross the road.

This is what the gameplay looks like. The player controls the turtle

which can only go forwards, and

there's a whole bunch of randomly generated cars

which are going horizontally across the screen. Now, once a player

reaches the other side of the screen, the cars speed up,

but the player goes back to the starting position,

ready to cross the screen again. Now,

at some point when the player turtle hits a car,

then that is game over.